+++ functions/_adapters/golemio/services/alerts/schemas.ts
@@ -22,13 +22,9 @@ export const pidRssItemSchema = z.object({
     }).optional().catch(null)
 });
 
-
-
 export const golemioRouteSchema = z.object({
     route_id: z.string(),
     route_short_name: z.string(),
     route_type: z.number(),
     route_color: z.string().optional()
 });
-
-